backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
cbor2==5.8.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
fastapi==0.128.0
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-genai==1.5.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
grpclib==0.4.9
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httplib2==0.31.2
hyperframe==6.1.0
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
modal==1.3.1
multidict==6.7.1
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.7.1
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-multipart==0.0.22
requests==2.32.5
rich==14.3.1
rsa==4.9.1
shellingham==1.5.4
starlette==0.50.0
synchronicity==0.11.1
toml==0.10.2
tqdm==4.67.1
typer==0.21.1
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
watchfiles==1.1.1
websockets==16.0
yarl==1.22.0
